funcion para ver si la ruta existe o no
